fix: restored home page and event page
